requirements.txt
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
tqdm
PyYAML
timm
pillow
opencv-python